Add Refresh token

+++ src/main/java/com/notistris/identityservice/dto/response/AuthenticationResponse.java
@@ -12,6 +12,6 @@ public class AuthenticationResponse {
 
     String token;
     String refresh_token;
-    Integer duration;
+    long duration;
 
 }
